controller/custom_views/examplary_views/views_requirements.txt
qrcode==7.3.1
requests==2.28.1
speedtest-cli==2.1.3
feedparser==6.0.10
